feat: configuration ui support (1/2) - script side
Implement basic support for the configuration UI. This is implemented in the following way: we inject a window.api object on whitelisted sites, which allows them to directly set the desired values.

This commit also adds separate entry points for dev and release builds, to allow whitelisting extra api client sites on dev (like localhost).

Test: `npx webpack --config .\webpack\dev.ts`